feat(#13/CC-45): build software IN-GUEST — the devcontainer is build-capable, not just bootable
The decisive acceptance criterion: "a real workload builds successfully in-guest (a
toolchain compiles a program to a runnable binary)." Witnessed for real.

- witness an_amd64_devcontainer_builds_a_program_in_guest: a stock linux/amd64
  devcontainer with a toolchain layer (a static linux-amd64 TinyCC + a freestanding
  source) boots on the x86-64 core; PID 1 runs `tcc -static -nostdlib /hello.c -o
  /hello` and then `/hello` — so the guest COMPILES a program inside the image and
  RUNS the binary it built (CC45-TCC-EXIT:0 → CC45-BUILT-IN-GUEST:42). A real C
  compiler runs on the hand-written x86-64 core; the devcontainer is genuinely usable
  for development, the Codespaces/Gitpod bar, not merely bootable.
- fixture: vv/artifacts/cc45/build/{tcc,hello.c} — a static linux-amd64 TinyCC (pinned
  mob a338258d, 0.9.28rc, built static so it runs in the libc-free busybox rootfs) +
  a freestanding program (its own _start, raw write+exit — the compile depends only on
  tcc). build.sh step [4/4] builds it reproducibly; cc45.sha256 + SOURCE.txt pin it;
  PROVENANCE.md records it. tar_gz_x packs the toolchain layer with +x.
- vv: the witness is wired into the gated cc45 suite; CC-44 + CC-45 move from the
  target label to the live-checked set in vv/run.sh (their suites are gated in
  vv/suites/, not targets).
- docs: both 10-Quality-Requirements catalogs name the build-in-guest, occupancy-
  streamed, deployed-init, and dogfood witnesses, and the build-capable claim.

+++ crates/holospaces/src/emulator.rs
@@ -5338,7 +5338,7 @@ mod tests {
 
         // The guest sees the full declared capacity…
         assert_eq!(
-            disk.len() as u64,
+            disk.len(),
             DISK_BYTES,
             "the disk reports its full declared 1 TiB capacity"
         );
